first iteration of comments
