+++ internal/app/upstream/client_test.go
@@ -31,6 +31,8 @@ func TestOpenStreamShouldReturnErrorForInvalidTypeUrl(t *testing.T) {
 
 	respCh, err = client.OpenStream(ctx, &v2.DiscoveryRequest{})
 	assert.NotNil(t, err)
+	err, ok := err.(*UnsupportedResourceError)
+	assert.True(t, ok)
 	assert.Nil(t, respCh)
 
 	cancel()
